Proyecto_Final/requirements.txt
kafka-python
pandas
pymongo